2008-12-02	Joel Sherrill <joel.sherrill@OARcorp.com>
	* libcsupport/src/ttyname.c: Correct prototype of ttyname_r() to use
	size_t for third parameter.

+++ cpukit/libcsupport/src/ttyname.c
@@ -59,9 +59,9 @@ static char ttyname_buf[sizeof (_PATH_DEV) + MAXNAMLEN] = _PATH_DEV;
  *  ttyname_r() - POSIX 1003.1b 4.7.2 - Demetermine Terminal Device Name
  */
 int ttyname_r(
-  int   fd,
-  char *name,
-  int   namesize
+  int     fd,
+  char   *name,
+  size_t  namesize
 )
 {
   struct stat sb;
